Access the node event api
To complete the various methods of accessing data in Angular2 I swapped out the file url and inserted the url to the node api created earlier. To get it to work I needed to add the Accept header in the request and several Allow headers on the server.

I'm still using Promises to work with the response. I'll need to upgrade the service to deal with Observables.