Add optional indicators to event model properties
To start off attempting to post data to the server, I want to make the properties that accept MongoId's in the form of arrays optional. This may not be a bad idea since the event will be built over time and not all at once. Therefore, some data, such as the list of courses, won't be available when the basic properties of the event are added. I can update the event later with the missing data.