Prepare the student list view
The first step is to retrieve the data from the server api. Once the data is listed on the page, I need to be able to select some students to be in the event. The students that are selected should be removed from the list of students to prevent confusion.