Add background color to active link on menu
I want to change the background color of the side nav menu link if it's the active link. Vincent was able to solve the problem by adding the css property to a div inside of the list item with the active link class. The final piece of the puzzle was placing the css in the styles.css file, rather than the component css file.