Add update event feature
I wanted to be able to use the same form to add a new event and to update an existing event. The classes, courses, and students will need to be handled in their respective components, but the name, location, and date should be added or updated in the event component.

When events are retrieved, the fields populated with Mongo ids are populated which makes it impossible to update the remaining fields. The solution is to use separate methods to return a list of events that are populated and a list which is not populated. This requires adding a series of parameters to the request url to retrieve the objects.