Refactor side nav menu
The routerLink directive works better when placed on the list item rather than the anchor tag. This allows the user to click anywhere in the list item making it easier to navigate using the menu. When only the anchor tag is used the cursor must be precisely on the text to find the active area. Since the background of the list item changes to indicate the active item it breaks the intuitive aspect of the link not to also allow the user to click in the item space too.