Refactor the list of students
The intention is to present the list of all students and the list of students in the selected event on the same page. However, I don't think the student component should be tied to the event components or the event service. So the goal is to use an event component to render the list on the student list component.