Implement Observable to access list of events
At this point I can make the request to getEvents which uses the EventService getEvents method to make the API call to the server. It gets back an Observable which is subcribed in the dashboard component.

By using the type, Observable<any>, it seems that I am able to successfully set the events property in the component to the events property on the response. Next, I need to see if I can simply set the events property in the event service before returning it to the dashboard component (as a standard array).

Ultimately, the components will never need to be concerned about whether the data is in the form of a Promise or an Observable. It just gets an array of Events which can be manipulated using the standard methods and directives available to arrays.