Add a callback to the getEvents method
The goal is to load the events into the events property on the event service, then access that property throughout the app when the events are needed.

Using the callback makes sure the http call ends an d the data is available before the dashboard loads and presents the list of events. There is likely a better solution, but this works for now.