Remove the event service calls
Initially I wanted to make the call to retrieve the list of events and simultaneously set the events property in the event service from the app component. However, by applying the rxjs methods on the Observable returned by the http GET method call in the event service and the dashboard component all the data binding issues seem to be resolved.