Clean up the list of events
The current list items contain an md-button class which creates a weird buton inside of the list items when the cursor hovers over an item. I Removed the button elements and replaced them with anchor tags. Using some CSS, I changed the colors and added a hover effect for the edit and remove links.