Track selected event by id
To prevent the chance that an event in the list of events might not be the same as the event of the same id in the selected event object, I want to simply pass and store the event id of the current event in the event service. Whenever additional properties of this event are needed, the id will be used to retrieve the event object from the list. All modifications to an event should update the event in the event list therefore preventing multiple versions of the same event floating around in the app.