Use async and slice pipes to list events
It seems that since the http request is returning an Observable, I can simply use the async pipe in the template to internally handle the subscription. Therefore, I can continue to pass around the Observable with no need to convert it to another form. Also, the slice pipe is an easy way to limit the number of items listed. Now I just need to figure out how to list additional events is the user wants to view more than I've rendered.