Handle selecting and deleting an event
When an event is selected from the list on the dashboard, the event property on the EventService is updated. The delete method simply removes the event from the database.

To help with the response types I added an Interface to make sure that the message property is enforced. Right now the data property is optional.