Change the student property type
Using a union property type that indicates the student property could be either an array of strings or an array of Student objects caused a type error attempting to apply array methods. It's an identified bug which will eventually be fixed. For now, the property type can be represented as an array of strings since I won't be populating the objects.