Add student lookup feature to add form
When the user attempts to add another student to the event three possibilities exist: the student does not exist in the db, the student already exists in the db, or the student info has been modified.

It is highly useful to autofill the form values if the student already exists. This can be accomplished using Observables.